Added a ResourceTracker system
If you right-click while dragging a box, it will attempt to buy wood
flooring (and save it) for the selected area.  This costs 5 wood per
square.

If you have enough in your storage, it will spend and construct the
flooring; otherwise it silently fails.

Also, there's a storage system for that now, as the ResourceTracker
class.

Also, Game1 has been renamed to CowMouseGame.  Sorry for the confusion.