Support for Buildings added
When you build a rectangle (currently, of flooring) it's designated as a
building, and other buildings can't be built on top of it.

Also, invalid selections are now marked, and if you try, it turns red
instead of yellow (this happens from overdoing the cost or overlapping
the other buildings).