Log Collectors
Added the "Log Hunter" class, as well as the "Log" class.  Now logs will
be randomly strewn about on the ground, and the NPC will wander around
and collect them.

Along the way, there was added some more intelligent path following
behavior (*not* pathfinding behavior, yet) which is probably keepable.

There was also some "resource" code (in the WorldManager class
specifically) that is absolutely a placeholder.