Added an FPS counter, began stress testing
For reasonable path sizes, it takes a tiny but noticable blip for a dude
to path to a thing, which obviously gets worse with more people
(linearly) and longer path sizes (quadratically).

A good solution would be to multi-thread the pathing; regardless of what
Toady says, this seems easy and natural.  I'll try to make this happen
tonight, along with dirty bits (so they can detect that a path is no
longer good, or even no longer optimal).