Walls now implemented; need some thread safety
Specifically, walls (as a generic non-passable building) are currently
in there.  Cool!  There are occasional crashes:

Situation:
If you add a wall to the buildings list while things are pathfinding, it
will sometimes crash.

Diagnosis:
The buildings queue is being altered (in the main thread) while it's
being iterated through (in the pathfinding thread) which is verboten.

Fix (to be implemented):
The buildings queue should be made threadsafe