[Minor] Minimum time (0.5s) between pathfinding
There was a problem with framerate; if it could be quickly determined
that there was no path from here to there, threads would be rapidly
started and stopped, which (it turns out) devastates framerates.  A
better solution will probably present itself eventually ("well, nothing
has changed, so don't bother trying right now") but this works for now.