Improved decision making and renamed some files.
Decision making is now smarter; they will not seek out a resource unless
there exists both a resource and a stockpile to put it in.  In fact they
won't even start that AI state.  Same thing for bed hunting.

In celebration of them becoming more of a real NPC, the following files
have been renamed and codified:

Townsman -> Person (now represents the physical body of a person, and
must be extended to implement Update)
LogHunter -> NPC (now represents a general purpose NPC)