[Major] Overhauled AI code
There is now a Goal class (with two specific instances, Goal_Carryable
and Goal_Zone) which allow for clean marking and unmarking for path
purposes.  It helps save a lot of code complexity and checking!

Also, startState, endOfPath, and getInterrupted are now dictionaries of
actions rather than annoying switch statements.