Beginning of AI revamp is done
The AI for NPCs is now completely based on Tasks, and following them, so
that they don't have to have so much customized code for each little
thing.  This is great, but I have to redo a lot of things.

I'm going to slowly integrate all the old functionality back into this
system, but that hasn't been done yet, so this is a side branch.