NPCs now respect walls, tasks can be canceled
Tasks (now FullTask and TaskStep, which I think is a better pair of
names) are now both abstract, where making extension classes are just
incredibly simple, with a few mostly easy abstract methods to implement
(involving cleanup from a canceled task, which can be canceled in the
middle).

From there, the wall thing is easy- if the path suddenly doesn't make
sense, it just cancels the task.  The WorldManager will automatically
reschedule it if it still makes sense.