[Bug Fix] Stockpiles can now be used >1 time
Surprisingly, this is a different bug than before, although I noticed it
at the same time.

Main issue: the Carryable "IsInStockpile" tag depends only on its
current state, which is wrong- when it's marked for collection, it could
still be in a stockpile.  So, you know.

Anyway, that's fixed (it's in a stockpile iff it has one in mind, now)